+++ src/app/accounts/platforms.ts
@@ -9,17 +9,19 @@ import * as actions from './actions';
 import store from '../store/store';
 import { loadingTracker } from '../shell/loading-tracker';
 import { goToLoginPage } from '../bungie-api/authenticated-fetch';
-import {
-  accountsSelector,
-  currentAccountSelector,
-  loadAccountsFromIndexedDB,
-  accountsLoadedSelector,
-} from './reducer';
+import { accountsSelector, currentAccountSelector, accountsLoadedSelector } from './selectors';
 import { ThunkResult } from 'app/store/types';
 import { dedupePromise } from 'app/utils/util';
 import { removeToken } from '../bungie-api/oauth-tokens';
 import { deleteDimApiToken } from 'app/dim-api/dim-api-helper';
-import { del } from 'idb-keyval';
+import { del, get } from 'idb-keyval';
+
+const loadAccountsFromIndexedDBAction: ThunkResult = dedupePromise(async (dispatch) => {
+  console.log('Load accounts from IDB');
+  const accounts = await get<DestinyAccount[] | undefined>('accounts');
+
+  dispatch(actions.loadFromIDB(accounts || []));
+});
 
 const getPlatformsAction: ThunkResult<readonly DestinyAccount[]> = dedupePromise(
   async (dispatch, getState) => {
@@ -31,7 +33,7 @@ const getPlatformsAction: ThunkResult<readonly DestinyAccount[]> = dedupePromise
 
     if (!getState().accounts.loadedFromIDB) {
       try {
-        await dispatch(loadAccountsFromIndexedDB());
+        await dispatch(loadAccountsFromIndexedDBAction);
       } catch (e) {
         console.error('Unable to load accounts from IDB', e);
       }
