+++ src/app/dim-api/actions.ts
@@ -8,7 +8,7 @@ import { RootState, ThunkResult } from 'app/store/types';
 import { DimApiState } from './reducer';
 import { get, set } from 'idb-keyval';
 import { getPlatforms } from '../accounts/platforms';
-import { currentAccountSelector } from '../accounts/reducer';
+import { currentAccountSelector } from 'app/accounts/selectors';
 import { observeStore } from '../utils/redux-utils';
 import _ from 'lodash';
 import {
