+++ src/app/search/FilterHelp.tsx
@@ -3,7 +3,7 @@ import './FilterHelp.scss';
 import React from 'react';
 import { RootState } from 'app/store/types';
 import { connect } from 'react-redux';
-import { destinyVersionSelector } from '../accounts/reducer';
+import { destinyVersionSelector } from 'app/accounts/selectors';
 import { t } from 'app/i18next-t';
 import { DestinyVersion } from '@destinyitemmanager/dim-api-types';
 import UserGuideLink from 'app/dim-ui/UserGuideLink';
